@@ -4,6 +4,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Text;
+using System.Text.Json;
 
 namespace ThinSdk.Neo.IO.Json
 {
@@ -55,7 +56,7 @@ namespace ThinSdk.Neo.IO.Json
 
         public override string AsString()
         {
-            return string.Join(VALUE_SEPARATOR.ToString(), items.Select(p => p?.AsString()));
+            return string.Join(",", items.Select(p => p?.AsString()));
         }
 
         public void Clear()
@@ -93,32 +94,6 @@ namespace ThinSdk.Neo.IO.Json
             items.Insert(index, item);
         }
 
-        internal new static JArray Parse(TextReader reader, int max_nest)
-        {
-            SkipSpace(reader);
-            if (reader.Read() != BEGIN_ARRAY) throw new FormatException();
-            JArray array = new JArray();
-            SkipSpace(reader);
-            if (reader.Peek() != END_ARRAY)
-            {
-                while (true)
-                {
-                    JObject obj = JObject.Parse(reader, max_nest - 1);
-                    array.items.Add(obj);
-                    SkipSpace(reader);
-                    char nextchar = (char)reader.Read();
-                    if (nextchar == VALUE_SEPARATOR) continue;
-                    if (nextchar == END_ARRAY) break;
-                    throw new FormatException();
-                }
-            }
-            else
-            {
-                reader.Read();
-            }
-            return array;
-        }
-
         public bool Remove(JObject item)
         {
             return items.Remove(item);
@@ -129,27 +104,34 @@ namespace ThinSdk.Neo.IO.Json
             items.RemoveAt(index);
         }
 
-        public override string ToString()
+        internal override void Write(Utf8JsonWriter writer)
         {
-            StringBuilder sb = new StringBuilder();
-            sb.Append(BEGIN_ARRAY);
+            writer.WriteStartArray();
             foreach (JObject item in items)
             {
-                if (item == null)
-                    sb.Append(LITERAL_NULL);
+                if (item is null)
+                    writer.WriteNullValue();
                 else
-                    sb.Append(item);
-                sb.Append(VALUE_SEPARATOR);
+                    item.Write(writer);
             }
-            if (items.Count == 0)
+            writer.WriteEndArray();
+        }
+
+        public override JObject Clone()
+        {
+            var cloned = new JArray();
+
+            foreach (JObject item in items)
             {
-                sb.Append(END_ARRAY);
+                cloned.Add(item.Clone());
             }
-            else
-            {
-                sb[sb.Length - 1] = END_ARRAY;
-            }
-            return sb.ToString();
+
+            return cloned;
+        }
+
+        public static implicit operator JArray(JObject[] value)
+        {
+            return new JArray(value);
         }
     }
 }
