@@ -3,6 +3,7 @@ using System.Globalization;
 using System.IO;
 using System.Text;
 using System.Text.Encodings.Web;
+using System.Text.Json;
 
 namespace ThinSdk.Neo.IO.Json
 {
@@ -31,50 +32,6 @@ namespace ThinSdk.Neo.IO.Json
             return Value;
         }
 
-        internal static JString Parse(TextReader reader)
-        {
-            SkipSpace(reader);
-            if (reader.Read() != QUOTATION_MARK) throw new FormatException();
-            char[] buffer = new char[4];
-            StringBuilder sb = new StringBuilder();
-            while (true)
-            {
-                int c = reader.Read();
-                if (c == QUOTATION_MARK) break;
-                if (c == '\\')
-                {
-                    c = (char)reader.Read();
-                    switch (c)
-                    {
-                        case QUOTATION_MARK: c = QUOTATION_MARK; break;
-                        case '\\': c = '\\'; break;
-                        case '/': c = '/'; break;
-                        case 'b': c = '\b'; break;
-                        case 'f': c = '\f'; break;
-                        case 'n': c = '\n'; break;
-                        case 'r': c = '\r'; break;
-                        case 't': c = '\t'; break;
-                        case 'u':
-                            reader.Read(buffer, 0, buffer.Length);
-                            c = short.Parse(new string(buffer), NumberStyles.HexNumber);
-                            break;
-                        default: throw new FormatException();
-                    }
-                }
-                else if (c < ' ' || c == -1)
-                {
-                    throw new FormatException();
-                }
-                sb.Append((char)c);
-            }
-            return new JString(sb.ToString());
-        }
-
-        public override string ToString()
-        {
-            return $"{QUOTATION_MARK}{JavaScriptEncoder.Default.Encode(Value)}{QUOTATION_MARK}";
-        }
-
         public override T TryGetEnum<T>(T defaultValue = default, bool ignoreCase = false)
         {
             try
@@ -85,6 +42,26 @@ namespace ThinSdk.Neo.IO.Json
             {
                 return defaultValue;
             }
+        }
+
+        internal override void Write(Utf8JsonWriter writer)
+        {
+            writer.WriteStringValue(Value);
+        }
+
+        public override JObject Clone()
+        {
+            return this;
+        }
+
+        public static implicit operator JString(Enum value)
+        {
+            return new JString(value.ToString());
+        }
+
+        public static implicit operator JString(string value)
+        {
+            return value == null ? null : new JString(value);
         }
     }
 }
