@@ -19,12 +19,12 @@ namespace ThinSdk.Neo.SmartContract.Manifest
         /// The methods field is an array containing a set of methods to be called. It can also be assigned with a wildcard *. If it is a wildcard *, then it means that any method can be called.
         /// If a contract invokes a contract or method that is not declared in the manifest at runtime, the invocation will fail.
         /// </summary>
-        public WildCardContainer<string> Methods { get; set; }
+        public WildcardContainer<string> Methods { get; set; }
 
         public static readonly ContractPermission DefaultPermission = new ContractPermission
         {
             Contract = ContractPermissionDescriptor.CreateWildcard(),
-            Methods = WildCardContainer<string>.CreateWildcard()
+            Methods = WildcardContainer<string>.CreateWildcard()
         };
 
         /// <summary>
@@ -37,7 +37,7 @@ namespace ThinSdk.Neo.SmartContract.Manifest
             return new ContractPermission
             {
                 Contract = ContractPermissionDescriptor.FromJson(json["contract"]),
-                Methods = WildCardContainer<string>.FromJson(json["methods"], u => u.AsString()),
+                Methods = WildcardContainer<string>.FromJson(json["methods"], u => u.AsString()),
             };
         }
 
